// src/modules/garmentsProduction/reports/cuttting/fabric-quality-problem-report/fabric-quality-problem-report-type.ts
export type FabricQualityProblemReportType = {
  PROBLEM_DATE: string | null;
  DEPARTMENT_NAME: string | null;
  BUYER: string | null;
  STYLENO: string | null;
  PONO: string | null;
  PROBLEM_KG: number | null;
  PROBLEM_PCS: number | null;
  KNITTING_HOUSE_ID: number | null;
  REMARKS: string | null;
  KNITTING_HOUSE_ID2: number | null;
  KNITTING_HOUSE_ID3: number | null;
  KNITTING_HOUSE_ID4: number | null;
  ID: number;
  KNITTING_HOUSE_NAME: string | null;
  KNITTING_HOUSE_NAME2: string | null;
  KNITTING_HOUSE_NAME3: string | null;
  KNITTING_HOUSE_NAME4: string | null;
  BATCH_NO: string | null;
  BATCH_ID: number | null;
  BATCH_QTY: number | null;
  FABRIC_NAME: string | null;
  COMPANY_NAME: string | null;
  COMPANY_ADDRESS: string | null;
  COMPANY_REMARKS: string | null;
  INTER_COMPANY_ID: number | null;
  BUYER_ID: number | null;
  PO_ID: number | null;
  STYLE_ID: number | null;
  DYEING_BATCH_NO: string | null;
  YARN_LOT_NUMBER: string | null;
  REQ_GSM: string | null;
  RCV_ROLL_QTY: number | null;
  RCV_QUANTITY: number | null;
  SUPPLIER: string | null;
  STITCH_LENGTH: string | null;
  BRAND_NAME: string | null;
  COLORNAME: string | null;
  YARN_COUNT: string | null;
  COMPANY_PREFIX: string | null;
};